chore(deps): Use forked record dependency for macOS fix
Override the `record` and `record_macos` packages to point to a
git repository instead of the published version from pub.dev.

This is a temporary measure to incorporate a yet to be merged fix for broken
echo cancellation on the OSX. The `master` branch
of the forked repository contains this necessary patch.

This override will be removed once the fix is included in an
official release of the record package.

Signed-off-by: Keyur Maru <kd2maru@gmail.com>